// src/api/generated/models/GetModelResultDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
import type { GetMaterialResultDto } from './GetMaterialResultDto';
export type GetModelResultDto = {
  _id: string;
  modelName: string;
  active: boolean;
  createBy: string;
  createAt: string;
  __v: number;
  materialId: GetMaterialResultDto;
};